requirements_all.txt
alembic==1.16.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
auto-logger==0.1.2
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
chess==1.11.2
click==8.1.8
colorlog==6.9.0
dacite==1.9.2
debugpy==1.8.14
defusedxml==0.7.1
fastjsonschema==2.21.1
Flask==3.0.3
Flask-Compress==1.17
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistune==3.1.3
mypy_extensions==1.1.0
narwhals==1.44.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.3.1
packaging==24.2
pandas==2.3.0
pandocfilters==1.5.1
pillow==10.4.0
platformdirs==4.3.8
pluggy==1.6.0
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.9
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.2
pytest==8.4.1
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-mock==3.14.1
python-chess==1.999
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.25.1
scipy==1.15.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.35
stchess @ git+https://github.com/TheoLvs/streamlit-chess@58a6171db5768b73b5caabf1d7ab4d96b8db43b9
stockfish==3.28.0
streamlit==1.39.0
tenacity==9.1.2
tinycss2==1.4.0
toml==0.10.2
toolz==1.0.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==5.0.3
webencodings==0.5.1
Werkzeug==3.1.3
zipp==3.22.0
zstandard==0.23.0
